Split Flood OSM building IF.
