+++ safe/gui/tools/test/test_wizard_dialog.py
@@ -1233,7 +1233,7 @@ class WizardDialogTest(unittest.TestCase):
         expected_flood_structure_functions_count = 3
         expected_raster_polygon_functions_count = 1
         expected_functions_count = 1
-        chosen_if = 'FloodBuildingImpactFunction'
+        chosen_if = 'FloodRasterBuildingImpactFunction'
 
         expected_hazard_layers_count = 1
         expected_exposure_layers_count = 1
@@ -1326,7 +1326,7 @@ class WizardDialogTest(unittest.TestCase):
         message = 'Expected flood impact function not found: %s' % chosen_if
         self.assertTrue(chosen_if in flood_ifs, message)
 
-        # step_fc_function: select FloodBuildingImpactFunction and press ok
+        # step_fc_function: select FloodRasterBuildingImpactFunction and press ok
         chosen_if_row = flood_ifs.index(chosen_if)
         dialog.lstFunctions.setCurrentRow(chosen_if_row)
         dialog.pbnNext.click()
