+++ safe/impact_functions/test/test_utilities.py
@@ -24,6 +24,8 @@ from safe.impact_functions.utilities import (
     keywords_to_str,
     add_to_list,
     get_python_file)
+from safe.impact_functions.inundation.flood_raster_OSM_building_impact import (
+    FloodRasterBuildingImpactFunction)
 
 
 class TestUtilities(unittest.TestCase):
@@ -71,7 +73,6 @@ class TestUtilities(unittest.TestCase):
         message = 'Expecting %s in %s' % (path, expected_paths)
 
         self.assertIn(path, expected_paths, message)
-
 
 if __name__ == '__main__':
     unittest.main()
